% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks_fun.R
\name{breaks_fun}
\alias{breaks_fun}
\title{Define x-axis break intervals based on chromosome size}
\usage{
breaks_fun(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
Returns a numeric vector with breaks between 0 and x.
}
\description{
This function defines break intervals on ggplot x-axis proportional to chromosome size.
If :
   - chromosome length < 100 Mb, set breaks every 10 Mb
   - chromosome length > 100 Mb but < 200 Mb, set breaks every 20 Mb
   - chromosome length > 200 Mb but < 350 Mb, set breaks every 50 Mb
   - chromosome length > 350 Mb, set breaks every 100 Mb.
}
