% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_redundant_regions.R
\name{concatReg}
\alias{concatReg}
\title{Concatenate redundant outlier genomic regions}
\usage{
concatReg(winList, rtab, listRegionWin, clustsize = 10, corr_thr = 0.6)
}
\arguments{
\item{rtab}{A data.frame containing the outlier genomic regions identified (generated by the function getReg)}

\item{clustsize}{Minimum size (in number of genomic windows) of the intersect portion between two overlapping genomic regions (default=10)}

\item{corr_thr}{Minimum threshold for correlation test (default=0.6)}

\item{tab}{A (4+n)-columns data.frame containing the following columns: genomic window IDs (ID), chromosome name (Chr), the start (posB) and end (posE) positions of the genomic window, and the centered-reducted coordinate values of the n MDS dimensions (MDScr1, MDScr2, MDScr3, ...)}

\item{listRegWin}{A n-levels list of the n identified outlier genomic regions by the function getReg, each containing nested list(s) of indexes of the genomic windows contained by these regions}
}
\value{
This function returns a list containing two elements : ("tabF") a table of concatenated outlier regions, ("to_remove") a list of all window IDs removed
}
\description{
Concatenate redundant outlier genomic regions
}
