% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_inversions.R
\name{detectInv}
\alias{detectInv}
\title{Identify putative inverted genomic regions from region-based PCA plots and k-means clustering}
\usage{
detectInv(
  vcf_in,
  tab_regions,
  popFile,
  outDir,
  bcftools,
  bgzip,
  minSize = 2,
  keep_tmp = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{vcf_in}{The VCF/BCF file used to conduct localPCA analysis (character string)}

\item{tab_regions}{R data.frame generated by getReg() and containing the identified outlier genomic regions}

\item{popFile}{Tabulated-text file containing 3 columns with the individual IDs (as in the VCF header), their assigned population (unique alphanumerical name, no space), and the Rcolor to be used for plotting, respectively. The Rcolor column is optional, only black dots will be shown on plots if missing.}

\item{outDir}{Output directory where to store per-region VCFs and associated PCA figures}

\item{bcftools}{By default, bcftools software should be installed in /usr/bin. If this is not the case, use this option to specify the user-defined full path to bcftools}

\item{bgzip}{By default, bgzip software should be installed in /usr/bin. If this is not the case, use this option to specify the user-defined full path to bgzip Required if keep_tmp=TRUE}

\item{minSize}{Minimum size of the outlier genomic regions (in Mb) to analyse. (default=2)}

\item{keep_tmp}{Whether to keep the per-region VCF files generated (may take up some space if large genome) (default=FALSE)}

\item{verbose}{Whether to display information on the analysis (default=TRUE)}
}
\description{
Identify putative inverted genomic regions from region-based PCA plots and k-means clustering
}
