% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outlier_regions.R
\name{getRegion}
\alias{getRegion}
\title{Define specific genomic regions harbouring population structure deviating from the genome-wide structure.}
\usage{
getRegion(
  winList,
  nMDS = 40,
  Gtest_thr = 0.05,
  min_nwin_H1 = 4,
  nwin_gap = 5,
  nwin_cluster = 3,
  concat_redundant = TRUE,
  clustsize = 10,
  corr_thr = 0.6
)
}
\arguments{
\item{winList}{Output list from \bold{detect_outlierWin}. Must contains 4-levels named: "coor.raw", "coor.cr", "pval" and "fdr"}

\item{nMDS}{Number of MDS dimension inferred, (integer, default=40)}

\item{Gtest_thr}{Minimum threshold for G-test significance (float, default=0.05)}

\item{min_nwin_H1}{Minimum number of genomic windows required to keep a chromosome for further analysis (integer, default=4)}

\item{nwin_gap}{Mininum number of non-significant genomic windows between two significant genomic windows (integer, default=3)}

\item{nwin_cluster}{Minimum number of genomic windows within a regions to define it as an "outlier region" potentially carrying strong within-population structure signal}

\item{concat_redundant}{Concatenate redundant overlapped genomic regions between MDS dimensions (boolean, default=TRUE)}

\item{clustsize}{Minimum size (in number of genomic windows) of the intersect portion between two overlapping genomic regions. Required if \option{concat_redundant=TRUE}. (integer, default=10)}

\item{corr_thr}{Minimum threshold for correlation tests between pairs of overlapped regions. Required if \option{concat_redundant=TRUE}. (default=0.6)}
}
\value{
If \option{concat_redundant=TRUE}, returns a list of 3 elements :
        \tabular{ll}{
          tabF \tab a data.frame of the concatenated outlier regions\cr
          tabF_raw \tab a data.frame of the "raw" outlier regions detected\cr
          listRegionWin \tab a list of regions with their respective window IDs encompassed
        }
        If \option{concat_redundant=TRUE}, returns a list of 2 elements (note here that tabF refers to the "raw" regions):
        \tabular{ll}{
          tabF \tab a data.frame of the "raw" outlier regions detected\cr
          listRegionWin \tab a list of regions with their respective window IDs encompassed
        }
}
\description{
Define specific genomic regions harbouring population structure deviating from the genome-wide structure.
}
