% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_datasets.R
\docType{data}
\name{invList}
\alias{invList}
\title{Example dataset of detectInv() result}
\format{
A list of 7 variables
}
\usage{
data(invList)
}
\description{
A list-type dataset containing sub-lists of 4 variables for each outlier regions analysed.
The variables within each sub-list are as follows:
}
\details{
\itemize{
  \item mat. Data frame with 126 obs. of 12 variables (sample.id, pop, pop.cols, group, cluster, group.cols, PC1, PC2, PC3, het.nsites, het.obs, het.exp)
  \item pca.var. Numerical vector [1:3] containing precentage of explained variance on the first three PCs
  \item het.wilcox.test. Rstatix_test class object containing result from wilcoxon test
  \item het.bonf.corr. Numeric list containing Bonferroni-corrected p-values for significance levels \*, \*\* and \*\*\*
}
}
\keyword{datasets}
