% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{regList}
\alias{regList}
\title{Example outlier regions obtained with getRegion()}
\format{
A list of 8 variables
}
\usage{
data(regList)
}
\description{
A list-type dataset containing 7 or 8 output variables from the function getRegion().
The output variables are as follows:
}
\details{
\itemize{
  \item wins. Data frame with 757 obs. of 4 variables
  \item coor.raw. Data frame with 757 obs. (genomic windows) of 40 variables (MDS) containing raw MDS coordinate values
  \item coor.cr. Data frame with 757 obs. (genomics windows) of 40 variables (MDS) containing centered-reducted MDS coordinate values
  \item pval. Data frame with 757 obs. (genomic windows) of 40 variables (MDS) containing p-values from normal distribution of coor.cr data
  \item fdr. Data frame with 757 obs. (genomic windows) of 40 variables (MDS) containing FDR rejection test result based on p-values data
  \item tabF_raw. Pre-final data frame with 12 obs. (outlier genomic regions) of 7 variables (ID, MDS, chr, start, end, noutliers, SizeRegions). Present only concat_redundant=TRUE, otherwise this data frame is stored as tabF
  \item tabF. Final data frame with 7 obs. (outlier genomic regions) of 8 variables (nID, ID, MDS, chr, start, end, noutliers, SizeRegions). Outlier regions are corrected for redundancy if concat_redundant=TRUE (here TRUE)
  \item listRegionWin. List of genomic window indices for each of the outlier genomic regions detected (here 12)
}
}
\keyword{datasets}
