% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_datasets.R
\docType{data}
\name{winList}
\alias{winList}
\title{Example dataset of loPCA() result}
\format{
A list of 7 variables
}
\usage{
data(winList)
}
\description{
A list-type dataset containing results from the 4-steps local PCA procedure from Li and Ralph (2019).
The variables are as follows:
}
\details{
\itemize{
  \item winfun. Function generated by lostruct::vcf_windower() to divide SNP data into windows of fixed size (here 100-SNPs windows)
  \item pcs_all. Matrix with 757 obs. (genomic windows) of 255 variables (PCA coordinates from first 2 PCs for each of the 126 samples) generated by lostruct::eigen_windows()
  \item rows_na. Indices of rows containing NA in pcs_all
  \item pcs. Same as pcs_all, but excluding rows_na (identical to pcs_all if rows_na=0)
  \item pcdist. Dissimilarity matrix of 757x757 obs. (numeric values) generated by lostruct::pc_dist()
  \item fit2d. List of 5 variables generated by lostruct::cmdscale() that summarize pcdist across 40 MDS
  \item data. Data frame with 757 obs. (genomic windows) of 44 variables
}
}
\keyword{datasets}
